requirements.txt
fastapi
jinja2
pandas
plotly
psycopg[binary]
python-dotenv
requests
uvicorn
